+++ packages/core/src/tracing/constants.js
@@ -1,10 +1,10 @@
 'use strict';
 
-exports.spanIdHeaderName = 'X-INSTANA-S';
-exports.spanIdHeaderNameLowerCase = exports.spanIdHeaderName.toLowerCase();
-
 exports.traceIdHeaderName = 'X-INSTANA-T';
 exports.traceIdHeaderNameLowerCase = exports.traceIdHeaderName.toLowerCase();
+
+exports.spanIdHeaderName = 'X-INSTANA-S';
+exports.spanIdHeaderNameLowerCase = exports.spanIdHeaderName.toLowerCase();
 
 exports.traceLevelHeaderName = 'X-INSTANA-L';
 exports.traceLevelHeaderNameLowerCase = exports.traceLevelHeaderName.toLowerCase();