+++ packages/google-cloud-run/test/using_api/app.js
@@ -9,8 +9,7 @@
 
 const http = require('http');
 
-const sendToParent = require('../../../serverless/test/util/send_to_parent');
-const { delay } = require('../../../core/test/test_util');
+const { delay, sendToParent} = require('../../../core/test/test_util');
 
 const instana = require('../..');